+++ src/Projectors/ProjectsEvents.php
@@ -14,16 +14,6 @@ trait ProjectsEvents
         return $this->name ?? get_class($this);
     }
 
-    /** @deprecated Use reset state instead */
-    public function reset(): void
-    {
-        if (! method_exists($this, 'resetState')) {
-            throw CouldNotResetProjector::doesNotHaveResetStateMethod($this);
-        }
-
-        $this->resetState();
-    }
-
     public function shouldBeCalledImmediately(): bool
     {
         return ! $this instanceof QueuedProjector;